serveless-embeddings/lambdas/code/requirements.txt
langchain
langchain-community
langchain-experimental
faiss-cpu
pypdf
boto3
pillow